% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annuity.R
\name{annuity}
\alias{annuity}
\title{Annuity of a credit loan}
\usage{
annuity(principal, rate, k, n)
}
\arguments{
\item{principal}{Amount of money originally invested or loaned, on which basis interest and returns are calculated.}

\item{rate}{Nominal interest rate (in percentages).}

\item{k}{Number of repayments per year.}

\item{n}{Total number of repayments/periods.}
}
\value{
The annuity (repayment amount in each period).
}
\description{
Annuity of a credit loan
}
\examples{
# Loan for 10000 EUR with a total of 12 monthly repayments and 6 \% annual interest rate.
annuity(10000, 6, 12, 12)
}
