% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atools_package.R
\docType{package}
\name{atools}
\alias{atools-package}
\alias{atools}
\title{Tools to Manipulate ATO}
\description{
Provides tools to manipulate ATO objects. Includes functions for filtering, excluding, and additional calculations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/trackyverse/atools}
  \item \url{http://atools.trackyverse.org}
  \item Report bugs at \url{https://github.com/trackyverse/atools/issues}
}

}
\author{
\strong{Maintainer}: Hugo Flávio \email{hflavio@dal.ca}

}
\keyword{internal}
