% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_below_threshold.R
\name{balance_below_threshold}
\alias{balance_below_threshold}
\title{Number of repayments before existing loan balance falls below a certain amount}
\usage{
balance_below_threshold(principal, rate, k, n, threshold)
}
\arguments{
\item{principal}{Amount of money originally invested or loaned, on which basis interest and returns are calculated.}

\item{rate}{Nominal interest rate (in percentages).}

\item{k}{Number of repayments per year.}

\item{n}{Total number of repayments/periods.}

\item{threshold}{Goal remaining loan amount.}
}
\value{
Number of repayments before existing loan balance falls below a certain amount
}
\description{
Number of repayments before existing loan balance falls below a certain amount
}
\examples{
balance_below_threshold(10000, 6, 12, 36, 5000)
}
\references{
Toman, Aleš (2016)
\emph{Rešene naloge iz finančne matematike 1, Finančni instrumenti.}
Ljubljana: DMFA založništvo.
}
