% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationMeasures.R
\name{binomLL}
\alias{binomLL}
\title{Binomial log likelihood}
\usage{
binomLL(y, y.hat)
}
\arguments{
\item{y}{status yes=1, no=0 or dead=1, alive=0}

\item{y.hat}{numeric. risk between 0 and 1}
}
\value{
log likelihood, numeric.
}
\description{
little wrapper around \code{\link{dbinom}}
}
\author{
Andi Boeck
}
