% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOSTmeta.R
\name{Brier}
\alias{Brier}
\title{Brier: Bierer score / MSPE for non-cencored outcomes}
\description{
  with no survival outcome, just the mean squared predicted error (MSPE). 
}
\usage{
  Brier(y, y.hat, scaled = FALSE)
}
\arguments{
  \item{y.hat}{numeric. risk between 0 and 1}
  
  \item{y}{status yes=1, no=0 or dead=1, alive=0}
  
  \item{scaled}{If TRUE scaled version is used.}
}
\Details{
  The scaling depends on the predictions y.hat and not only on the actual outcome y as in NagR2. This limits the use of the scaled version to assess different models on external data. 
}
\value{
  Brier score numeric.
}
