% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{calc_det_interval}
\alias{calc_det_interval}
\title{Calculate time interval between detections}
\usage{
calc_det_interval(object)
}
\arguments{
\item{object}{an \code{\link[ATO]{ATO}}}
}
\value{
an updated \code{\link[ATO]{ATO}}
}
\description{
Calculate time interval between detections
}
