% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_error.R
\name{check_error}
\alias{check_error}
\title{Check if evaluation of expression returns any errors}
\usage{
check_error(expr)
}
\arguments{
\item{expr}{A R expression}
}
\value{
Boolean. TRUE if evaluating the expression returns any errors and FALSE otherwise.
}
\description{
Check if evaluation of expression returns any errors
}
\examples{
x <- 3
check_error(3 * x)
rm(x)
check_error(3 * x)
}
