% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_warning.R
\name{check_warning}
\alias{check_warning}
\title{Check if evaluation of expression returns any warnings}
\usage{
check_warning(expr)
}
\arguments{
\item{expr}{A R expression}
}
\value{
Boolean. TRUE if evaluating the expression returns any warnings and FALSE otherwise.
}
\description{
Check if evaluation of expression returns any warnings
}
\examples{
x <- 3
check_warning(3 * x)
check_warning(as.numeric("3a"))
}
