% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationMeasures.R
\name{cInt}
\alias{cInt}
\title{Calibration intercept}
\usage{
cInt(y, y.hat)
}
\arguments{
\item{y}{status yes=1, no=0 or dead=1, alive=0}

\item{y.hat}{numeric. risk between 0 and 1}
}
\value{
numeric.
}
\description{
Intercept of regreestion with slope fixed to one, 
measures calibration-in-the-large
}
\author{
Andi Boeck
}
