% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_project_structure.R
\name{copy_project_structure}
\alias{copy_project_structure}
\title{Title}
\usage{
copy_project_structure(
  path = getwd(),
  folder_name = "New project",
  overwrite = FALSE
)
}
\arguments{
\item{path}{Where to create the new project. Defaults to current working directory.}

\item{folder_name}{Name of the new folder. Defaults to "New project".}

\item{overwrite}{Whether to overwrite an existing folder with the same name as \emph{folder_name}}
}
\value{
Nothing.
}
\description{
Title
}
\examples{
\dontrun{
copy_project_structure()
}
}
