% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{create_filter_vec}
\alias{create_filter_vec}
\title{Worker function to create filtering vector}
\usage{
create_filter_vec(object, slt = c("ani", "dep", "det", "tag", "obs"), checks)
}
\arguments{
\item{object}{an ATO}

\item{slt}{the ATO slot to filter}

\item{checks}{a list of column names and respective values to filter with}
}
\value{
a logical vector
}
\description{
Worker function to create filtering vector
}
