% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_functions.R
\name{.create_filter_vec}
\alias{.create_filter_vec}
\title{worker function to create filtering vector}
\usage{
.create_filter_vec(object, checks, slt = c("det", "dep", "tag", "ani", "obs"))
}
\arguments{
\item{object}{an ATO}

\item{checks}{a list of column names and respective values to filter with}

\item{slt}{the ATO slot to filter}
}
\value{
a logical vector
}
\description{
worker function to create filtering vector
}
\keyword{internal}
