% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_functions.R
\name{.dyn_round}
\alias{.dyn_round}
\title{Round value to a reasonable number of decimal places}
\usage{
.dyn_round(value, max = 10)
}
\arguments{
\item{value}{the value to be rounded}

\item{max}{the maximum number of decimal places allowed}
}
\value{
the rounded value
}
\description{
Round value to a reasonable number of decimal places
}
\keyword{internal}
