% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{dummy_ints}
\alias{dummy_ints}
\title{Create dummy numeric data}
\usage{
dummy_ints(cols, rows, limit = 10, replacement = TRUE, seed = NULL)
}
\description{
Create a tibble of random numeric data
}
\details{
Supports a maximum of 702 columns and a random seed
}
