% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_functions.R
\name{exclude_orphan_dets}
\alias{exclude_orphan_dets}
\title{Wrapper to exclude detections that to not match the deployments}
\usage{
exclude_orphan_dets(object, hard = FALSE)
}
\arguments{
\item{object}{an \code{\link[ATO]{ATO}}}

\item{hard}{If false (the default), detections are flagged as invalid through
the invalid column, but kept in the dataset. Switch to true to completely
remove the detections from the dataset (useful to improve performance when
handling very large datasets).}
}
\value{
the updated ATO
}
\description{
Wrapper to exclude detections that to not match the deployments
}
