% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filter}
\alias{filter}
\title{Filter rows from an ATO slot}
\usage{
filter(
  object,
  slt = c("ani", "dep", "det", "tag", "obs"),
  ...,
  hard = FALSE,
  silent = FALSE
)
}
\arguments{
\item{object}{an \code{\link[ATO]{ATO}}}

\item{slt}{The slot to be filtered.}

\item{...}{named arguments to filter for. The argument's name must match
one of the column names in the animals slot. If only one value is provided,
data is filtered by exact match. If two values are provided for continuous
variables, data is filtered within the interval provided. If three or more
values are provided, data is filtered by exact match.}

\item{hard}{If false (the default), animal rows are flagged as invalid
through the valid column, but kept in the dataset. Switch to true to drop
the tags that do not fit the filtering criteria (useful to improve
performance when handling very large datasets).}

\item{silent}{Logical: Supress summary messages}
}
\value{
the updated ATO
}
\description{
Filter rows from an ATO slot
}
