% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filter_datetime}
\alias{filter_datetime}
\title{Wrapper to filter detections and observations within the desired time period}
\usage{
filter_datetime(object, from = "1970-01-01", to = "3000-01-01", hard = FALSE)
}
\arguments{
\item{object}{an \code{\link[ATO]{ATO}}}

\item{from}{Filter data from this timestamp (included). Can be posixct or character.}

\item{to}{Filter data up to this timestamp (included). Can be posixct or character.}

\item{hard}{If false (the default), detections are flagged as invalid through
the invalid column, but kept in the dataset. Switch to true to drop
detections that do not fit the filtering criteria (useful to improve
performance when handling very large datasets).}
}
\value{
the updated ATO
}
\description{
Wrapper to filter detections and observations within the desired time period
}
