% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filter_dep}
\alias{filter_dep}
\title{Filter rows from the @dep slot}
\usage{
filter_dep(object, ...)
}
\arguments{
\item{object}{an \code{\link[ATO]{ATO}}}

\item{...}{named arguments to filter for. The argument's name must match
one of the column names in the deployments If only one value is provided,
data is filtered by exact match. If two values are provided for continuous
variables, data is filtered within the interval provided. If three or more
values are provided, data is filtered by exact match.}
}
\value{
the updated ATO
}
\description{
Filter rows from the @dep slot
}
