% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filter_det}
\alias{filter_det}
\title{Filter rows from the @det slot}
\usage{
filter_det(object, ..., hard = FALSE)
}
\arguments{
\item{object}{an \code{\link[ATO]{ATO}}}

\item{...}{named arguments to filter for. The argument's name must match
one of the column names in the detections. If only one value is provided,
data is filtered by exact match. If two values are provided for continuous
variables, data is filtered within the interval provided. If three or more
values are provided, data is filtered by exact match.}

\item{hard}{If false (the default), detections are flagged as invalid through
the invalid column, but kept in the dataset. Switch to true to drop
detections that do not fit the filtering criteria (useful to improve
performance when handling very large datasets).}
}
\value{
the updated ATO
}
\description{
Filter rows from the @det slot
}
