% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filter_ping_dev}
\alias{filter_ping_dev}
\title{wrapper to filter detections by ping deviation bands}
\usage{
filter_ping_dev(object, bands, grace, hard = FALSE)
}
\arguments{
\item{object}{an \code{\link[ATO]{ATO}}}

\item{bands}{a vector of deviation bands to filter}

\item{grace}{width of the bands to extract}

\item{hard}{If false (the default), detections are flagged as invalid through
the invalid column, but kept in the dataset. Switch to true to drop
detections that do not fit the filtering criteria (useful to improve
performance when handling very large datasets).}
}
\value{
the updated ATO
}
\description{
wrapper to filter detections by ping deviation bands
}
