% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filter_reset}
\alias{filter_reset}
\title{Clear all previously applied (non-hard) filters}
\usage{
filter_reset(object, s)
}
\arguments{
\item{object}{an \code{\link[ATO]{ATO}}}

\item{s}{the slot to be reset. If not provided, all slots are reset.}
}
\value{
the updated ATO
}
\description{
Clear all previously applied (non-hard) filters
}
