% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filter_targeted}
\alias{filter_targeted}
\title{Wrapper to filter detections and observations that match the
transmitters in the tag slot.}
\usage{
filter_targeted(object, hard = FALSE)
}
\arguments{
\item{object}{an \code{\link[ATO]{ATO}}}

\item{hard}{If false (the default), detections are flagged as invalid through
the invalid column, but kept in the dataset. Switch to true to drop
detections that do not fit the filtering criteria (useful to improve
performance when handling very large datasets).}
}
\value{
the updated ATO
}
\description{
Wrapper to filter detections and observations that match the
transmitters in the tag slot.
}
