% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_delim.R
\name{fmt_delim}
\alias{fmt_delim}
\title{Format delimeter}
\usage{
fmt_delim(data_loc = "clipboard", delim_before = " ", delim_after = ", ")
}
\arguments{
\item{data_loc}{Path to where the data is stored. Defaults to clipboard.}

\item{delim_before}{Delimeter used in the supplied string.}

\item{delim_after}{Replacement of the original delimeter.}
}
\value{
A string with the new delimeter.
}
\description{
Useful when copying a list with space or other characters as delimeter.
}
\examples{
\dontrun{
fmt_delim()
}
}
