% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_num.R
\name{fmt_num}
\alias{fmt_num}
\title{Format numeric values.}
\usage{
fmt_num(x, decimal_mark = ",", big_mark = NULL)
}
\arguments{
\item{x}{Vector with numeric values.}

\item{decimal_mark}{Decimal separator. Defaults to ","}

\item{big_mark}{Separator between digits of large numbers.}
}
\value{
A formatted character vector.
}
\description{
Format numeric values.
}
\examples{
fmt_num(50000.54, decimal_mark = ",")
fmt_num("50000.54", decimal_mark = ",")
}
