% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_str.R
\name{fmt_str}
\alias{fmt_str}
\title{Format string}
\usage{
fmt_str(
  str,
  ascii = TRUE,
  u_case = FALSE,
  rm_special = FALSE,
  rm_space = FALSE,
  n_char = NULL
)
}
\arguments{
\item{str}{A character vector}

\item{ascii}{Whether to transliterate characters to ASCII. This means removal of carons and similiar.}

\item{u_case}{Whether to return upper case letters.}

\item{rm_special}{Whether to remove special characters.}

\item{rm_space}{Whether to remove spaces.}

\item{n_char}{Maximum number of characters in each element of the vector.}
}
\value{
A character vector.
}
\description{
Format string
}
\examples{
fmt_str("tek\%kl oaf", u_case = FALSE, rm_special = TRUE, n_char = 5)
}
