% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{group_sum}
\alias{group_sum}
\title{Single Var. Group By and Summarize}
\usage{
group_sum(data, group, method, col)
}
\description{
Helper function that groups by a single variable and applies a basic summary
to one column
}
\details{
I do this too often to justify continuing to write group_by and summarize
}
