% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationMeasures.R
\name{Hosmer}
\alias{Hosmer}
\title{Hosmer-Lemeshow statistic}
\usage{
Hosmer(y, y.hat)
}
\arguments{
\item{y}{status yes=1, no=0 or dead=1, alive=0}

\item{y.hat}{numeric. risk between 0 and 1}
}
\value{
numeric.
}
\description{
Groups by deciles.
}
\details{
Chi-Square typ measure.
}
\author{
Andi Boeck
}
