% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationMeasures.R
\name{Hosmer2}
\alias{Hosmer2}
\title{Hosmer-Lemeshow statistic}
\usage{
Hosmer2(y, y.hat, g = 10)
}
\arguments{
\item{y}{status yes=1, no=0 or dead=1, alive=0}

\item{y.hat}{numeric. risk between 0 and 1}

\item{g}{how many groups.}
}
\value{
numeric.
}
\description{
Differt version to \code{Hosmer}.
}
\details{
Chi-Square typ measure.
}
\author{
Andi Boeck
}
