% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lom.R
\name{lom}
\alias{lom}
\title{Last day of month}
\usage{
lom(date)
}
\arguments{
\item{date}{A vector of date objects or objects that can be coerced to date}
}
\value{
A vector of dates.
}
\description{
Last day of month
}
\examples{
dates <- seq.Date(from = as.Date("2022-01-15"), to = as.Date("2022-12-15"), by = "1 month")
lom(dates)
}
