% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{more_letters}
\alias{more_letters}
\title{Extend built-in letter vector}
\usage{
more_letters(upper = FALSE)
}
\description{
Stupid little utility to extend the default letter vector past 26 (e.g.
'AA', 'AB' ...)
}
