% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationMeasures.R
\name{NagR2}
\alias{NagR2}
\title{Nagelkerke R-square measure}
\usage{
NagR2(y, y.hat)
}
\arguments{
\item{y}{status yes=1, no=0 or dead=1, alive=0}

\item{y.hat}{numeric. risk between 0 and 1}
}
\value{
R-square, standardized version.
}
\description{
likelihood based R-square
}
\author{
Andi Boeck
}
