% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{Ncon}
\alias{Ncon}
\title{Number of concordant pairs}
\usage{
Ncon(y, y.hat)
}
\arguments{
\item{y}{status yes=1, no=0 or dead=1, alive=0}

\item{y.hat}{numeric. risk between 0 and 1}
}
\description{
Concordant pairs
}
