\name{Nrel}
\alias{Nrel}
\title{Number of relevant pairs}
\usage{
Nrel(y)
}
\arguments{
  \item{y}{status yes=1, no=0 or dead=1, alive=0}
}
\description{
Relevant pairs
}
\examples{
outcome <- c(1,1,0,0);
 Nrel(outcome)
}
\references{
Tutz, Analyse kategorialer Daten, p. 111ff
}

