% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationMeasures.R
\name{oisse}
\alias{oisse}
\title{Combine different validation measures in \code{data.frame}.}
\usage{
oisse(y, y.hat)
}
\arguments{
\item{y}{status yes=1, no=0 or dead=1, alive=0}

\item{y.hat}{numeric. risk between 0 and 1}
}
\value{
data.frame with several measures as columns
}
\description{
Combine different validation measures in \code{data.frame}.
}
\examples{
outcome <- c(1,1,0,0); marker <- c(.3, .2, .1,.8);
oisse(marker, outcome)
}
\author{
Andi Boeck
}
