% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pars}
\alias{pars}
\title{General parameters}
\format{
A named list object with elements representing parameters for a certain task or possible values of an attribute.

\describe{
\item{palette_cb}{Color-blind friendly palette in hexadecimal notation. Source: http://www.cookbook-r.com/Graphs/Colors_(ggplot2)}
}
}
\source{
Own
data("pars)
}
\usage{
pars
}
\description{
Lisst containing predefined parameters.
}
\keyword{datasets}
