% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loan_balance.R
\name{plot_loan_balance}
\alias{plot_loan_balance}
\title{Plot outstanding principal balance over time}
\usage{
plot_loan_balance(principal, rate, k, n)
}
\arguments{
\item{principal}{Amount of money originally invested or loaned, on which basis interest and returns are calculated.}

\item{rate}{Nominal interest rate (in percentages).}

\item{k}{Number of repayments per year.}

\item{n}{Total number of repayments/periods.}
}
\value{
The outstanding principal balance of a loan (existing loan).
}
\description{
Plot outstanding principal balance over time
}
\examples{
loan_balance(10000, 6, 6, 12, 12)
}
\references{
Toman, Aleš (2016)
\emph{Rešene naloge iz finančne matematike 1, Finančni instrumenti.}
Ljubljana: DMFA založništvo.
}
