% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spearman.ci.R
\name{print.spearman.ci}
\alias{print.spearman.ci}
\title{Print method for speraman.ci}
\usage{
##S3 method for class \code{spearman.ci}.
}
\arguments{
\item{x}{Object of class \code{spearman.ci}.}

\item{digits}{Digits to show.}

\item{...}{Not used.}
}
\description{
Print method for speraman.ci
}
