% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_caron.R
\name{remove_caron}
\alias{remove_caron}
\title{Remove carons}
\usage{
remove_caron(str)
}
\arguments{
\item{str}{A character vector}
}
\value{
A character vector
}
\description{
Remove carons
}
\examples{
\dontrun{
remove_caron("string_with_carons")
}
}
