% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_ws.R
\name{remove_ws.character}
\alias{remove_ws.character}
\title{Remove whitespace from character vector.}
\usage{
\method{remove_ws}{character}(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector without white space.
}
\description{
Remove whitespace from character vector.
}
\examples{
\dontrun{
x <- c("     Hi", "Hey    ", "  Hello   ")
remove_ws(x)
}
}
