% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_ws.R
\name{remove_ws.data.frame}
\alias{remove_ws.data.frame}
\title{Remove whitespace from character columns.}
\usage{
\method{remove_ws}{data.frame}(x)
}
\arguments{
\item{x}{A data.frame.}
}
\value{
A data.frame without white space in character columns.
}
\description{
Remove whitespace from character columns.
}
\examples{
\dontrun{
x <- data.frame(test = c("     Hi", "Hey    ", "  Hello   "))
remove_ws(x)
}
}
