% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_ws.R
\name{remove_ws}
\alias{remove_ws}
\title{Remove whitespace from character columns}
\usage{
remove_ws(dt)
}
\arguments{
\item{dt}{A data.table object}
}
\value{
Nothing. Removal is done by assignment.
}
\description{
Remove whitespace from character columns
}
\examples{
\dontrun{
x <- data.table(test = c("     Hi", "Hey    ", "  Hello   "))
remove_ws(x) # Nothing is returned.
x$test
}
}
