% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationMeasures.R
\name{ROCdat}
\alias{ROCdat}
\title{Prepare \code{data.frame} for ROC-curve}
\usage{
ROCdat(y.hat, y)
}
\arguments{
\item{y.hat}{numeric. risk between 0 and 1}

\item{y}{status yes=1, no=0 or dead=1, alive=0}
}
\value{
\code{data.frame}
}
\description{
Cutoff with best Youden index
}
\details{
If low values of predictor/risk/\code{y.hat} correspond to event of
 interest coded with \code{y=1}, this will result in an ROC-curve below
 the diagonal.
}
\examples{
 n <- 1e2
 y.hat <- seq(0, 1, length=n) 
 y <- rbinom(n, size = 1, prob = y.hat) 
 rocdat <- ROCdat(y.hat = y.hat, y = y)
 
 with(rocdat, plot(x= 1 - Specificity, y = Sensitivity, type = "l"))
 (a <- Ycutoff(y.hat=y.hat, y))
 (p <- rocdat[which(rocdat$thres == a),])
 points(x=1-p$Spec, y=p$Sens, col="red")
 
 # reverse direction
 y <- rbinom(n, size = 1, prob = 1- y.hat) 
 rocdat <- ROCdat(y.hat = y.hat, y = y)
 with(rocdat, plot(x = 1 - Specificity, y = Sensitivity, type = "l"))
}
\author{
Andi Boeck
}
