% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_wd.R
\name{set_wd}
\alias{set_wd}
\title{Set working directory}
\usage{
set_wd(..., file_loc = FALSE)
}
\arguments{
\item{...}{String containing directory.}

\item{file_loc}{Whether to set working directory to document path.}
}
\value{
Nothing.
}
\description{
Set working directory
}
\examples{
#set_wd()
}
