% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_lat_to_ascii.R
\name{str_lat_to_ascii}
\alias{str_lat_to_ascii}
\title{Transliterate characters from Latin to ASCII}
\usage{
str_lat_to_ascii(str)
}
\arguments{
\item{str}{A character vector}
}
\value{
A character vector in UTF-8/ASCII encoding.
}
\description{
Transliterate characters from Latin to ASCII
}
\examples{
\dontrun{
str_lat_to_ascii("string_with_carons")
}
}
