% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toLatexMat.R
\name{toLatexMat}
\alias{toLatexMat}
\title{Write matrix to console to copy paste in latex}
\usage{
toLatexMat(mat, digits = 2)
}
\arguments{
\item{mat}{matrix}

\item{digits}{round to those digits. Defaults to 2.}
}
\description{
Just prints in console
}
\examples{
 mat <- matrix(rnorm(100), 10,10)
 toLatexMat(mat)
}
\author{
Andi Boeck
}
