% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{TPC}
\alias{TPC}
\title{True Postive Count}
\usage{
TPC(thres, y, y.hat)
}
\arguments{
\item{thres}{thershold where to split. Must be in range of \code{y.hat}}

\item{y}{status yes=1, no=0 or dead=1, alive=0}

\item{y.hat}{numeric. risk between 0 and 1}
}
\value{
true positive count
}
\description{
Area under the ROC curve. Calculated in various ways.
}
\author{
Andi Boeck
}
