% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_pars.R
\name{update_pars}
\alias{update_pars}
\title{Convenience function used to update the general parameters data}
\usage{
update_pars()
}
\value{
List of parameters.
}
\description{
Convenience function used to update the general parameters data
}
\examples{
update_pars()
}
