% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_vsc_snippets.R
\name{update_vsc_snippets}
\alias{update_vsc_snippets}
\title{Update VS Code snippets}
\usage{
update_vsc_snippets(
  from = "C:\\\\Users\\\\andre\\\\AppData\\\\Roaming\\\\Code\\\\User\\\\snippets",
  to = "C:/Users/andre/OneDrive/AS/PROJECTS/atools/inst/vscode_snippets"
)
}
\arguments{
\item{from}{Where to copy the files from}

\item{to}{Copy the files to}
}
\value{
"Success!"
}
\description{
Update VS Code snippets
}
\examples{
\dontrun{
update_vsc_snippets()
}
}
