% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationMeasures.R
\name{WilcoxHand}
\alias{WilcoxHand}
\title{Wilcoxon statistic by hand}
\usage{
WilcoxHand(y, y.hat)
}
\arguments{
\item{y}{status yes=1, no=0 or dead=1, alive=0}

\item{y.hat}{numeric. risk between 0 and 1}
}
\value{
numeric.
}
\description{
Wilcoxon statistic by hand.
}
\details{
Just to know how R does calculate it.
}
\author{
Andi Boeck
}
