% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_str.R
\name{wrap_str}
\alias{wrap_str}
\title{Wrap Character Strings to Format Paragraphs}
\usage{
wrap_str(x, col_len = 10)
}
\arguments{
\item{x}{A character vector or an object which can be converted to a character vector by as.character}

\item{col_len}{Maximum length of the column}
}
\value{
A character vector.
}
\description{
Wrap Character Strings to Format Paragraphs
}
\examples{
wrap_str("This line is too long to display in one row, hence it will be displayed in multiple rows", col_len = 10)
}
