% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_rows.R
\name{bind_rows}
\alias{bind_rows}
\title{Combine two objects by rows.}
\usage{
bind_rows(...)
}
\arguments{
\item{...}{Add}
}
\description{
Combine two objects by rows.
}
\examples{
\dontrun{
bind_rows(...)
}
}
