\name{cLarge}
\alias{cLarge}
\title{Calibration in-the-large}
\usage{
cLarge(y, y.hat)
}
\arguments{
  \item{y.hat}{numeric. risk between 0 and 1}

  \item{y}{status yes=1, no=0 or dead=1, alive=0}
}
\value{
numeric.
}
\description{
Calibration in-the-large
}
\details{
Mean prediction minus mean outcome.
}
\author{
Andi Boeck
}

