% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_functions.R
\name{.apply_filter}
\alias{.apply_filter}
\title{Worker function to apply filtering}
\usage{
.apply_filter(
  object,
  new_valid,
  type = c("det", "dep", "tag", "ani", "obs"),
  hard = FALSE
)
}
\arguments{
\item{object}{an ATO}

\item{new_valid}{a logical vector, normally created by
\code{\link{.create_filter_vec}}.}

\item{type}{the ATO slot to filter}

\item{hard}{Should the filtering mark rows as invalid (FALSE) or remove them
altogether (TRUE)}
}
\value{
the updated ATO
}
\description{
Worker function to apply filtering
}
\keyword{internal}
