% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_path.R
\name{fmt_path}
\alias{fmt_path}
\title{Format path to be readable by R}
\usage{
fmt_path(path = "clipboard", fwd_slash = FALSE, write = TRUE)
}
\arguments{
\item{path}{Path. Can be quoted or unquoted.}

\item{fwd_slash}{Whether to return single forward slashes. Defaults to FALSE, which returns double backslashes.}

\item{write}{Whether to write the new path to clipboard.}
}
\value{
A string containing the new path.
}
\description{
Format path to be readable by R
}
\examples{
x <- "C:\ example"
fmt_path(x, write = FALSE)
}
