% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{install_packages}
\alias{install_packages}
\title{Install custom packages list and / or update existing packages}
\usage{
install_packages(update = FALSE)
}
\arguments{
\item{update}{Whether to update installed packages with newer versions. Defaults to FALSE.}
}
\value{
Nothing.
}
\description{
Install custom packages list and / or update existing packages
}
\examples{
\dontrun{
install_packages()
}
}
