\name{NagR2}
\alias{NagR2}
\title{Nagelkerke R-square measure}
\usage{
NagR2(y, y.hat)
}
\arguments{
  \item{y.hat}{numeric. risk between 0 and 1}

  \item{y}{status yes=1, no=0 or dead=1, alive=0}
}
\value{
R-square, standardized version.
}
\description{
likelihood based R-square
}
\author{
Andi Boeck
}

