\name{print.spearman.ci}
\alias{print.spearman.ci}
\title{Print method for speraman.ci}
\usage{
##S3 method for class \code{spearman.ci}.
}
\arguments{
  \item{x}{Object of class \code{spearman.ci}.}

  \item{digits}{Digits to show.}

  \item{...}{Not used.}
}
\description{
Print method for speraman.ci
}

