% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SQL.R
\name{read_sql}
\alias{read_sql}
\title{Read a SQL file into a string in R.}
\usage{
read_sql(path, rm_comments = FALSE)
}
\arguments{
\item{path}{Path to the .sql file. If the sql file includes multiple queries or has other text before and after the query, mark the start of the query with "-- start query" and the end of the query with "-- end query".}

\item{rm_comments}{Whether to remove comments in sql code.}
}
\value{
A string containing the sql code.
}
\description{
Read a SQL file into a string in R.
}
\examples{
\dontrun{
 read_SQL("~file.sql")
}
}
