% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SQL.R
\name{read_SQL}
\alias{read_SQL}
\title{Read a SQL file into a string in R.}
\usage{
read_SQL(path)
}
\arguments{
\item{path}{Path to the .sql file.}
}
\value{
A string containing the sql code.
}
\description{
Read a SQL file into a string in R.
}
\examples{
\dontrun{
 read_SQL("~file.sql")
}
}
