% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_ws.R
\name{remove_ws}
\alias{remove_ws}
\title{Remove whitespace from character vectors/columns.}
\usage{
remove_ws(...)
}
\arguments{
\item{...}{Add}
}
\value{
The form of the value returned by summary depends on the class of its argument. See the documentation of the particular methods for details of what is produced by that method.
}
\description{
Remove whitespace from character vectors/columns.
}
\examples{
\dontrun{
x <- data.table(test = c("     Hi", "Hey    ", "  Hello   "))
remove_ws(x) # Nothing is returned.
x$test
}
}
