% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_ws.R
\name{remove_ws}
\alias{remove_ws}
\title{Remove whitespace from character vectors/columns.}
\usage{
remove_ws(...)
}
\arguments{
\item{...}{Add}

\item{x}{Any of the following: A character vector, a data.frame or a data.table.}
}
\value{
Nothing, if x is a data.table and same object as x without white space in character columns otherwise.
}
\description{
Remove whitespace from character vectors/columns.
}
\examples{
\dontrun{
x <- data.table(test = c("     Hi", "Hey    ", "  Hello   "))
remove_ws(x) # Nothing is returned.
x$test
}
}
